Move template info to its own module
